GSYE-757: Moved heatpump energy parameters to their own package. Fixed pylintrc configuration.
