+++ tests/strategies/energy_parameters/test_heat_pump.py
@@ -7,7 +7,7 @@ from gsy_framework.constants_limits import GlobalConfig, TIME_ZONE
 from gsy_framework.utils import generate_market_slot_list
 from pendulum import duration, today
 
-from gsy_e.models.strategy.energy_parameters.heat_pump import (
+from gsy_e.models.strategy.energy_parameters.heatpump.heat_pump import (
     HeatPumpEnergyParameters,
     TankParameters,
 )
